+++ lib/src/cars/data/datasource/remoteds.dart
@@ -34,6 +34,8 @@ class CarsRemoteDatasourceImpl implements CarsRemoteDatasource {
 
     if (response.statusCode != 200) throw Exception(decodedResponse);
 
-    return decodedResponse.map<CarsModel>((car) => CarsModel.fromJSON(car));
+    return decodedResponse
+        .map<CarsModel>((car) => CarsModel.fromJSON(car))
+        .toList();
   }
 }
