+++ lib/src/mainSection/cars/presentation/available_cars.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
 
-import 'package:rent_wheels/src/mainSection/cars/data/available_cars_data.dart';
-
-import 'package:rent_wheels/core/enums/enums.dart';
 import 'package:rent_wheels/core/widgets/sizes/sizes.dart';
 import 'package:rent_wheels/core/widgets/theme/colors.dart';
 import 'package:rent_wheels/core/global/globals.dart' as global;
@@ -25,9 +22,9 @@ class _AvailableCarsState extends State<AvailableCars> {
         child: Padding(
           padding:
               EdgeInsets.symmetric(horizontal: Sizes().width(context, 0.04)),
-          child: const AvailableCarsData(
-            type: AvailableCarsType.allAvailableCars,
-          ),
+          // child: const AvailableCarsData(
+          //   type: AvailableCarsType.allAvailableCars,
+          // ),
         ),
       ),
     );
